-fix: made price integer

+++ src/main/java/org/gamein/marketservergamein2022/core/dto/result/ProductDTO.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 public class ProductDTO {
     private final Long id;
     private final String name;
-    private final Double price;
+    private final Integer price;
     private final Integer level;
     private final Integer unitVolume;
     private final Long productionRate;
